+++ src/main/java/org/conteudos/Conteudo.java
@@ -3,27 +3,32 @@ package org.conteudos;
 public abstract class Conteudo {
     protected  int id;
     protected  String titulo;
-    protected  int duracao;
+    protected  double duracao;
 
-    public Conteudo(int id, String titulo, int duracao) {
+    public Conteudo(int id, String titulo, double duracao) {
         this.id = id;
         this.titulo = titulo;
         this.duracao = duracao;
     }
 
     public int getId(){
+
         return id;
     }
     public String getTitulo(){
+
         return titulo;
     }
-    public int getDuracao(){
+    public double getDuracao(){
+
         return duracao;
     }
     public void setDuracao(int duracao){
+
         this.duracao = duracao;
     }
     public void setTitulo(String titulo){
+
         this.titulo = titulo;
     }
