+++ src/main/java/org/conteudos/Musica.java
@@ -8,17 +8,19 @@ public class Musica extends Conteudo{
     private Artista artista;
     private Album album;
 
-    public Musica(int id, String titulo, int duracao, Artista artista, Album album){
+    public Musica(int id, String titulo, double duracao, Artista artista, Album album){
         super(id, titulo, duracao);
         this.artista = artista;
         this.album = album;
     }
 
     public void tocar(){
+
         System.out.println("Tocando música: " + titulo + " | Artista: " + artista.getNome());
     }
 
     public Artista getArtista() {
+
         return artista;
     }
     public Album getAlbum() {
